Circle functionality for new A-G scale
Updated circle functionality. Now it responds to each click instead of just dragging.  The correct angles have been set for the new A-G scale circle, and a radius feature has been added so it only registers a click when clicking in the circle.  Turned off the continuous line renderer and added a line renderer that draws straight lines between clicked notes.  Added a placeholder music trigger script and object that has a delegate called when the user clicks a note.

TODO: Right now each note can only be clicked once; they need to be able to be clicked multiple times.  But not sure what that should look like.  ALSO the line renderer looks poopy right now, that needs some love.